+++ src/helpers/analyzeDocument.ts
